AppRestaurante/GestionVehiculos/requirements.txt
argon2-cffi
argon2-cffi-bindings
asgiref
bootstrap4
cffi
crispy-bootstrap4
Django
django-cors-headers
django-crispy-forms
django-rest-framework
djangorestframework
djangorestframework_simplejwt
pillow
pycparser
PyJWT
sqlparse
tzdata
psycopg2-binary
django-imagekit
